empty criterion and neighborhood generator are defined
